feat(backend): Update image only after data is updated
New logic (completely rewrite) of how and when Images are updated in
/playerstats route and /groups route
When a data file is updated, check if the time difference between the
time modification of image and the data is greater than15 minutes and
that data was most recently updated than image, if that's the case then
generate a new image and upload to cloudinary..!!
Much better aproach then periodical updates. Even when data has not
changed.