feat: update table models to add their own row_id as primary key column
Update the database table models for the following tables:
- Hosts
- Matches
- PlayerStansings &
- TeamStandings
Remove `season_no` as column from all these models as it's useless (it
was an oversight during planning, this no wouldn't help in identifying
season based data)
Add own `row_id` to these models to uniqely identify each row across the
table.
Also, add `tournament_id` as another column to identify data based on a
particular season. (Not adding tournament slug & season as the
identifier as it's not a good choice)
Migration has been done on the database as necessary.