feat: add from_json & to_json methods to tournaments model (utility methods)
Add utility methods to easily construct a new data object row or to get
json_data from an existing object row. These are:
from_json (@classmethod) - call this to construct and return a new data
object row
to_json - call this on an existing data object that is tracking the row
to return it's full json in desired response. (in sqlalchemy)