fix: run image updation/deletion process in background while updating group standings in backend
In backend, update the put_groups callback to run image updation &
deletion processes in the background using threading, and return the
response early. This avoids waiting for all images to be updated and not
cause frontend client to be put on hold & cause a server error occured.
(because of a max 10 second limit on edge functions of cloud hosting.)
Additionally, refactor the put_groups callback code.
(refactor & improve put / edit standings logic in backend).