feat: show maximum 16 teams in standings page embed image
In standings page's embed image, cap the maximum teams that would be
shown on an embed image to 16. This is to prevent the image from being
too cluttered and hard to read. (for groups with more than 16 teams, like 24 or 32)
Modify PUT /groups backend endpoint to limit the number of teams to 16.